requirements.txt
Flask>=2.3
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
python-dotenv>=1.0
pytest>=7.0
